Support building images in custom repo
We have partial support for building images with a custom docker repo
by simply setting a REPO env var on the build host. This will build the
right images, but the environment variables that are set in the images
to control what agent gets deployed do not reflect the custom repo.

Fixed by passing REPO into docker build as a build arg and respecting it
in the Dockerfiles